# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth": {
      "get": {
        "operationId": "AuthController_getUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthToken"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthToken"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSessionAndUserResult"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/get-id": {
      "get": {
        "operationId": "AuthController_getUserId",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/property": {
      "post": {
        "operationId": "PropertyController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatePropertyDto"
                }
              }
            }
          }
        },
        "tags": [
          "Property"
        ]
      }
    },
    "/property/{id}": {
      "get": {
        "operationId": "PropertyController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Property"
        ]
      }
    },
    "/property/owner/{ownerId}": {
      "get": {
        "operationId": "PropertyController_getByOwner",
        "parameters": [
          {
            "name": "ownerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Property"
        ]
      }
    }
  },
  "info": {
    "title": "Domio",
    "description": "The domio property application",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Domio",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "SignInData": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthToken": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          }
        },
        "required": [
          "access_token"
        ]
      },
      "SignupDto": {
        "type": "object",
        "properties": {}
      },
      "IGetSession": {
        "type": "object",
        "properties": {
          "sessionToken": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "expires": {
            "type": "string"
          }
        },
        "required": [
          "sessionToken",
          "userId",
          "expires"
        ]
      },
      "GetSessionRole": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "description",
          "id",
          "createdAt"
        ]
      },
      "GetSessionUserWithRoles": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "format": "date-time",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetSessionRole"
            }
          }
        },
        "required": [
          "id",
          "email",
          "emailVerified",
          "name",
          "image",
          "roles"
        ]
      },
      "GetSessionAndUserResult": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/IGetSession"
          },
          "user": {
            "$ref": "#/components/schemas/GetSessionUserWithRoles"
          }
        },
        "required": [
          "session",
          "user"
        ]
      },
      "CreatePropertyDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "house",
              "apartment",
              "townhouse",
              "condo",
              "duplex",
              "commercial",
              "land"
            ]
          },
          "name": {
            "type": "string"
          },
          "streetName": {
            "type": "string"
          },
          "streetNumber": {
            "type": "number"
          },
          "suburb": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "bedrooms": {
            "type": "number"
          },
          "bathrooms": {
            "type": "number"
          },
          "hasParking": {
            "type": "boolean"
          },
          "isFurnished": {
            "type": "boolean"
          },
          "ownerId": {
            "type": "string"
          },
          "isForRent": {
            "type": "boolean"
          },
          "isForSale": {
            "type": "boolean"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "postalCode": {
            "type": "string"
          },
          "squareFeet": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "mainImage": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "name",
          "streetName",
          "streetNumber",
          "suburb",
          "country",
          "bedrooms",
          "bathrooms",
          "hasParking",
          "isFurnished",
          "ownerId",
          "isForRent",
          "isForSale",
          "images",
          "isActive"
        ]
      }
    }
  }
}